+++ BlueprintUI/Sources/Element/ElementContent.swift
@@ -3,7 +3,7 @@ import UIKit
 /// Represents the content of an element.
 public struct ElementContent {
 
-    private let storage: ContentStorage
+    let storage: ContentStorage
 
     //
     // MARK: Initialization
@@ -140,10 +140,13 @@ extension ElementContent {
     ///
     /// - parameter measurable: How to measure the `ElementContent`.
     public init(measurable: Measurable) {
-        self = ElementContent(
-            layout: MeasurableLayout(measurable: measurable),
-            configure: { _ in }
-        )
+//        self = ElementContent(
+//            layout: MeasurableLayout(measurable: measurable),
+//            configure: { _ in }
+//        )
+        storage = MeasurableStorage(measurer: { constraint, environment in
+            measurable.measure(in: constraint)
+        })
     }
 
     /// Initializes a new `ElementContent` with no children that delegates to the provided measure function.
@@ -208,7 +211,7 @@ extension ElementContent {
 }
 
 
-fileprivate protocol ContentStorage: SPContentStorage {
+protocol ContentStorage: SPContentStorage {
     var childCount: Int { get }
 
     func measure(
@@ -268,7 +271,7 @@ extension ElementContent {
         public var layout: LayoutType
 
         /// Child elements.
-        fileprivate var children: [Child] = []
+        var children: [Child] = []
 
         init(layout: LayoutType) {
             self.layout = layout
@@ -391,7 +394,7 @@ extension ElementContent {
             }
         }
 
-        fileprivate struct Child {
+        struct Child {
 
             var traits: LayoutType.Traits
             var key: AnyHashable?
@@ -402,121 +405,6 @@ extension ElementContent {
     }
 }
 
-extension ElementContent: Sizable {
-    func sizeThatFits(proposal: ProposedViewSize, environment: Environment) -> CGSize {
-        storage.sizeThatFits(proposal: proposal, environment: environment)
-    }
-}
-
-
-extension ElementContent.Builder {
-    func sizeThatFits(proposal: ProposedViewSize, environment: Environment) -> CGSize {
-        let subviews = children.map { child in
-            LayoutSubview(
-                element: child.element,
-                content: child.content,
-                environment: environment,
-                key: GenericLayoutValueKey<LayoutType>.self,
-                value: child.traits
-            )
-        }
-        return self.layout.sizeThatFits(proposal: proposal, subviews: subviews)
-    }
-    
-    
-    func performSinglePassLayout(context: SPLayoutContext, environment: Environment) -> [IdentifiedNode] {
-        
-//        let proposal = ProposedViewSize(attributes.bounds.size)
-        let attributes = context.attributes
-        let proposal = context.proposal
-        let frame = context.attributes.frame
-
-        let subviews = children.map { child in
-            LayoutSubview(
-                element: child.element,
-                content: child.content,
-                environment: environment,
-                key: GenericLayoutValueKey<LayoutType>.self,
-                value: child.traits
-            )
-        }
-        
-//        let bounds = CGRect(origin: .zero, size: attributes.bounds.size)
-        layout.placeSubviews(
-            in: frame,
-            proposal: proposal,
-            subviews: subviews
-        )
-        
-        var identifierFactory = ElementIdentifier.Factory(elementCount: children.count)
-
-        let identifiedNodes: [IdentifiedNode] = zip(children, subviews).map { (child, subview) in
-
-            let placement = subview.placement
-                ?? .init(position: attributes.center, anchor: .center, size: .proposal(proposal))
-
-            let size: CGSize
-            if let width = placement.size.width, let height = placement.size.height {
-                size = .init(width: width, height: height)
-//                print("\(type(of: subview.element)) placed at fixed size \(size)")
-            } else {
-                let measuredSize = subview.sizeThatFits(placement.size.proposal)
-                size = .init(
-                    width: placement.size.width ?? measuredSize.width,
-                    height: placement.size.height ?? measuredSize.height
-                )
-//                print("\(type(of: subview.element)) placed at measured \(measuredSize) and resolved to \(size)")
-            }
-            let childOrigin = placement.origin(for: size)
-
-
-
-                size: size
-            )
-            let offsetFrame = CGRect(
-                origin: childOrigin - frame.origin,
-                size: size
-            )
-//            if (frame.origin != .zero) {
-//                print("origin not zero")
-//            }
-//            if (childOrigin != .zero){
-//                print("child origin not zero: \(childOrigin)")
-//            }
-            print("\(type(of: subview.element)) frame \(childFrame) within \(frame)")
-
-            let childAttributes = LayoutAttributes(frame: offsetFrame)
-
-
-
-
-
-
-            let identifier = identifierFactory.nextIdentifier(
-                for: type(of: subview.element),
-                key: child.key
-            )
-            
-            let childContext = SPLayoutContext(
-                proposal: placement.size.proposal,
-                attributes: LayoutAttributes(frame: offsetFrame)
-            )
-
-            let node = LayoutResultNode(
-                element: child.element,
-                layoutAttributes: childAttributes,
-                environment: environment,
-                children: child.content.performSinglePassLayout(
-                    context: childContext,
-                    environment: environment
-                )
-            )
-            print("\(type(of: child.element)) result \(node.layoutAttributes.frame)")
-            return (identifier: identifier, node: node)
-        }
-        return identifiedNodes
-    }
-}
 
 private struct EnvironmentAdaptingStorage: ContentStorage {
     let childCount = 1
@@ -648,7 +536,7 @@ private struct MeasurableStorage: ContentStorage {
         return measurer(constraint, environment)
     }
 
-
+    func performSinglePassLayout(context: SPLayoutContext, environment: Environment) -> [IdentifiedNode] {
         []
     }
 }
@@ -702,32 +590,6 @@ fileprivate struct SingleChildLayoutHost: Layout {
         precondition(subviews.count == 1)
         return wrapped.placeSubview(in: bounds, proposal: proposal, subview: subviews[0])
     }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
 
 struct Measurer: Measurable {
